% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Surv.R
\name{export_is.Surv}
\alias{export_is.Surv}
\alias{is.Surv}
\title{\code{is.Surv} function imported from \code{survival}}
\description{
This function is imported from the \code{survival} package. See
\code{\link[survival]{is.Surv}}.
}
