\name{aftgee-package}
\alias{aftgee-package}
%\alias{aftgee}
\docType{package}
\title{
  Accelerate Failure Time with Generalize Estimating Equations
}
\description{
  This package features both rank-based approach and least squares
  approach to fit accelerated failure time model.
  For the rank-based approach, the package allows various weight choices
  and uses an induced smoothing procedure to improve computational
  efficiency.
  With the rank-based estimator as an initial value, the generalized
  estimating equation approach is used as an extension of the least
  squares approach to the multivariate case.
  Additional sampling weights are incorporated to handle missing data
  needed as in case-cohort studies.
  }
\details{
  \tabular{ll}{
    Package: aftgee
    Type: Package
    Version: 0.4-3
    Date: 2014-04-13
    License: GPL(>=3)
    LazyLoad: yes
  }
}
\author{
  Sy Han Chiou \email{schiou@d.umn.edu},
  Jun Yan \email{jun.yan@uconn.edu},
  Sangwook Kang \email{sangwook.kang@uconn.edu}
}
\references{
  Chiou, S. H. and Kang, S. and Yan, J. (2013)
  Fast Accelerated Failure Time Modeling for Case-Cohort Data.
  \emph{Statistics and Computing}, forthcoming.

  Chiou, S. H. and Kim, H. and Yan, J. (2014) Semiparametric
  Multivariate Accelerated Failure Time Model with Generalized
  Estimating Equations. \emph{Lifetime Data Analysis}, forthcoming.

  Jin, Z. and Lin, D. Y. and Ying, Z. (2006) On Least-squares Regression
  with Censored Data. \emph{Biometrika}, 90, 341--353.
}
\keyword{ aftgee }
