\name{aftgee-package}
\alias{aftgee-package}
%\alias{aftgee}
\docType{package}
\title{
  Accelerate Failure Time with Generalize Estimating Equations
}
\description{
  This package features both rank-based estimates and least square
  estimates to the Accelerated Failure Time (AFT) model. 
  For rank-based estimation, it provides approaches that include the
  computatonally efficent Gehan's weight and the general's weight such
  as the logrank weight. 
  For the least square estimation, the estimating equation is solved
  with Generalized Estimating Equations (GEE). 
  Moreover, in multivariate cases, the dependence working correlation
  strcture can be specified in GEE's setting. 
}
\details{
  \tabular{ll}{
    Package: aftgee
    Type: Package
    Version: 0.2-31
    Date: 2012-06-05
    License: GPL(>=3)
    LazyLoad: yes
  }
}
\author{
  Sy Han Chiou \email{steven.chiou@uconn.edu}, Jun Yan
  \email{jun.yan@uconn.edu}, Sangwook Kang \email{sangwook.kang@uconn.edu}
}
\references{
  Chiou, S. H. and Kim, H. and Yan, J. (2012) Semiparametric
  Multivariate Accelerated Failure Time Model with Generalized
  Estimating Equations. University of Connecticut. %% report number?
  
  Jin, Z. and Lin, D. Y. and Ying, Z. (2006) On Least-squares Regression
  with Censored Data. Biometrika, 90, 341--353.
  
  Johnson, L. M. and Strawderman, R. L. (2009) Induced Smoothing for the
  Semiparametric Accelerated Failure Time Model: Asymptotics and
  Extensions to Clustered Data. Biometrika, 96, 577--590.
}
\keyword{ aftgee }
